requirements.txt
# ── Web framework ──────────────────────────────────────────────
fastapi==0.111.0
uvicorn==0.29.0

# ── Database ────────────────────────────────────────────────────
psycopg2-binary==2.9.9
sqlalchemy==2.0.30

# ── Kafka streaming ─────────────────────────────────────────────
kafka-python==2.0.2

# ── Machine learning ────────────────────────────────────────────
xgboost==2.0.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

# ── Scheduling (perishable checker job) ─────────────────────────
schedule==1.2.1

# ── Utilities ───────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
requests==2.31.0